Fix copy-paste bug, signed/unsigned warnings, and yaml-cpp transitive dependency
- Fix copy-paste bug in ndarray_group_stream.hh: yfilenames.as<>() was
  missing [i] index when constructing patterns with path_prefix
- Fix ~30 signed/unsigned comparison warnings across 7 headers:
  ndarray_base.hh, lattice.hh, lattice_partitioner.hh, synthetic.hh,
  ndarray_stream.hh, ndarray_stream_vtk.hh, ndarray_group_stream.hh
- Replace deprecated reshape(coords) with reshapef(coords.shapef()) in
  synthetic.hh
- Make yaml-cpp a PRIVATE dependency of libndarray so downstream apps
  that don't use the stream API are not forced to find/link yaml-cpp

+++ include/ndarray/ndarray_base.hh
@@ -447,7 +447,7 @@ inline void ndarray_base::reshapec(const std::vector<size_t>& dims_)
 inline void ndarray_base::reshapec(const std::vector<int>& dims)
 {
   std::vector<size_t> mydimsc;
-  for (int i = 0; i < dims.size(); i ++)
+  for (size_t i = 0; i < dims.size(); i ++)
     mydimsc.push_back(dims[i]);
   reshapec(mydimsc);
 }
@@ -461,7 +461,7 @@ inline void ndarray_base::reshapec(size_t ndims, const size_t dims[])
 inline void ndarray_base::reshapef(const std::vector<int>& dims)
 {
   std::vector<size_t> mydims;
-  for (int i = 0; i < dims.size(); i ++)
+  for (size_t i = 0; i < dims.size(); i ++)
     mydims.push_back(dims[i]);
   reshapef(mydims);
 }
@@ -737,7 +737,7 @@ inline void ndarray_base::read_netcdf(int ncid, int varid, int ndims, const size
       NC_SAFE_CALL( nc_get_att_float(ncid, varid, "_FillValue", &fillvalue) );
 
       float *data = (float*)pdata();
-      for (int i = 0; i < nelem(); i ++)
+      for (size_t i = 0; i < nelem(); i ++)
         if (data[i] == fillvalue)
           data[i] = std::nan("");
     }
